refactor(storage-config): split StorageParams out from databend-common-meta-app
