+++ src/meta/app-storage/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "databend-common-meta-app-storage"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+publish.workspace = true
+edition.workspace = true
+
+[dependencies]
+databend-common-base = { workspace = true }
+databend-common-exception = { workspace = true }
+opendal = { workspace = true }
+serde = { workspace = true }
+
+[lints]
+workspace = true
