+++ src/meta/app-storage/src/lib.rs
@@ -12,5 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use databend_common_meta_app_storage;
-pub use databend_common_meta_app_storage::*;
+#![feature(no_sanitize)]
+mod storage_params;
+
+pub use storage_params::*;
