+++ src/meta/app/Cargo.toml
@@ -13,10 +13,10 @@ storage-hdfs = []
 anyerror = { workspace = true }
 chrono = { workspace = true }
 databend-common-ast = { workspace = true }
-databend-common-base = { workspace = true }
 databend-common-exception = { workspace = true }
 databend-common-expression = { workspace = true }
 databend-common-io = { workspace = true }
+databend-common-meta-app-storage = { workspace = true }
 databend-common-meta-app-types = { workspace = true }
 databend-common-meta-kvapi = { workspace = true }
 databend-common-meta-types = { workspace = true }
@@ -28,7 +28,6 @@ itertools = { workspace = true }
 maplit = { workspace = true }
 num-derive = { workspace = true }
 num-traits = { workspace = true }
-opendal = { workspace = true }
 paste = { workspace = true }
 prost = { workspace = true }
 serde = { workspace = true }
